+++ src/Kronika/Date/DayOfMonth.php
@@ -141,18 +141,6 @@ final readonly class DayOfMonth implements DateUnit
         return $this->compareTo($other)->lessOrEqual();
     }
 
-    /** @deprecated {@see self::is()} */
-    public function isEqualTo(self $other): bool
-    {
-        return $this->is($other);
-    }
-
-    /** @deprecated {@see self::isNot()} */
-    public function isNotEqualTo(self $other): bool
-    {
-        return $this->isNot($other);
-    }
-
     /**
      * Checks if this day of month is after or equal to another one.
      *