+++ src/Kronika/Date/DayOfWeek.php
@@ -219,18 +219,6 @@ enum DayOfWeek: int implements DateUnit
         return $this->compareTo($other)->lessOrEqual();
     }
 
-    /** @deprecated {@see self::is()} */
-    public function isEqualTo(self $other): bool
-    {
-        return $this->is($other);
-    }
-
-    /** @deprecated {@see self::isNot()} */
-    public function isNotEqualTo(self $other): bool
-    {
-        return $this->isNot($other);
-    }
-
     /**
      * Checks if this day of week is after or equal to another one.
      *