+++ src/Kronika/Date/Month.php
@@ -246,18 +246,6 @@ enum Month: int implements DateUnit
         return $this->compareTo($other)->lessOrEqual();
     }
 
-    /** @deprecated {@see self::is()} */
-    public function isEqualTo(self $other): bool
-    {
-        return $this->is($other);
-    }
-
-    /** @deprecated {@see self::isNot()} */
-    public function isNotEqualTo(self $other): bool
-    {
-        return $this->isNot($other);
-    }
-
     /**
      * Checks if this month is after or equal to another one.
      *