+++ src/Kronika/ZonedDateTime.php
@@ -433,18 +433,6 @@ final class ZonedDateTime extends \DateTimeImmutable implements DateTime
         return $this->compareTo($other, $precision)->lessOrEqual();
     }
 
-    /** @deprecated {@see self::is()} */
-    public function isEqualTo(DateTime|Unit|Native $other, Precision $precision = Precision::Micro): bool
-    {
-        return $this->is($other, $precision);
-    }
-
-    /** @deprecated {@see self::isNot()} */
-    public function isNotEqualTo(DateTime|Unit|Native $other, Precision $precision = Precision::Micro): bool
-    {
-        return $this->isNot($other, $precision);
-    }
-
     #[\Override]
     public function isAfterOrEqualTo(DateTime|Unit|Native $other, Precision $precision = Precision::Micro): bool
     {
@@ -481,18 +469,6 @@ final class ZonedDateTime extends \DateTimeImmutable implements DateTime
     public function toLocalDateTime(): LocalDateTime
     {
         return $this->local;
-    }
-
-    /** @deprecated {@see self::startOfMonth()} */
-    public function toStartOfMonth(): self
-    {
-        return $this->startOfMonth();
-    }
-
-    /** @deprecated {@see self::endOfMonth()} */
-    public function toEndOfMonth(): self
-    {
-        return $this->endOfMonth();
     }
 
     #[\Override]